Add idea about snapshots
Padraig pointed out that task sets being live-linked to classes means
that if you, say, have multiple years of classes in a file, tweaking the
task set means you're changing old marks.

Two possibilities here:

- treat the task sets as a template and copy the weightings to a given
  class, meaning weighting tweaks happen at the class level and we have
  isolation.
- have a place to hive off a snapshot of the inputs to the class mark
  calculations, ie. a cache. Have a moment where you decide to cache the
  marks, eg. after report cards are submitted.

The second is intriguing since it would also enable the possibility of
"what if" play on the weightings. You could snapshot the marks at any
time, not just end of term and not just when the task set changes, but
possibly any input. We should talk to Carole about whether this is even
interesting. But I've added a table of snapshots to capture the thought.